--- src/main/java/com/testbird/artisan/columnar/annotation/RowKey.java
@@ -1,21 +1,0 @@
-/**
- * 
- */
-package com.testbird.artisan.columnar.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-/**
- * mark this field is rowkey
- * 
- * @author zachary.zhang
- *
- */
-@Target(ElementType.FIELD)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface RowKey {
-
-}
