--- src/main/java/zachary/zhanghao/columnar/funciton/ColumnarAdmin.java
@@ -1,15 +1,0 @@
-/**
- * 
- */
-package zachary.zhanghao.columnar.funciton;
-
-import zachary.zhanghao.columnar.funciton.base.Schema;
-import zachary.zhanghao.columnar.funciton.base.Source;
-
-/**
- * @author zachary.zhang
- *
- */
-public interface ColumnarAdmin extends Schema, Source {
-
-}
