--- src/main/java/zachary/zhanghao/columnar/funciton/ColumnarClient.java
@@ -1,16 +1,0 @@
-/**
- * 
- */
-package zachary.zhanghao.columnar.funciton;
-
-import zachary.zhanghao.columnar.funciton.base.Aggregator;
-import zachary.zhanghao.columnar.funciton.base.Persistent;
-import zachary.zhanghao.columnar.funciton.base.Source;
-
-/**
- * @author zachary.zhang
- *
- */
-public interface ColumnarClient extends Aggregator, Persistent, Source {
-
-}
