--- src/main/java/zachary/zhanghao/columnar/funciton/base/Source.java
@@ -1,20 +1,0 @@
-/**
- * 
- */
-package zachary.zhanghao.columnar.funciton.base;
-
-import zachary.zhanghao.columnar.hbase.HBaseSource;
-
-/**
- * 提供对HBaseSource的操作功能
- * 
- * @author zachary.zhang
- *
- */
-public interface Source {
-
-    HBaseSource getHBaseSource();
-
-    void setHBaseSource(HBaseSource source);
-
-}
