--- src/test/java/com/testbird/artisan/hbaseclient/demo/Teacher.java
@@ -1,13 +1,0 @@
-/**
- * 
- */
-package com.testbird.artisan.hbaseclient.demo;
-
-/**
- * @author zachary.zhang
- *
- */
-public class Teacher {
-
-}
-
