+++ src/test/java/zachary/zhanghao/columnar/demo/Demo.java
@@ -9,8 +9,6 @@ import org.apache.hadoop.hbase.util.Bytes;
 
 import zachary.zhanghao.columnar.config.DataSourceConfig;
 import zachary.zhanghao.columnar.exception.ColumnarClientException;
-import zachary.zhanghao.columnar.funciton.ColumnarAdmin;
-import zachary.zhanghao.columnar.funciton.ColumnarClient;
 import zachary.zhanghao.columnar.hbase.HBaseColumnarAdmin;
 import zachary.zhanghao.columnar.hbase.HBaseColumnarClient;
 import zachary.zhanghao.columnar.hbase.HBaseSource;
@@ -25,7 +23,7 @@ public class Demo {
 
     public static void main(String[] args) {
         try {
-            ColumnarClient client = constructClient();
+            HBaseColumnarClient client = constructClient();
 
             putObject(client);
 
@@ -40,12 +38,12 @@ public class Demo {
             long count = client.count(null, null, User.class);
             System.out.println(count + " rows data in user table");
 
-            long sum = client.countAndSum(null, null, User.class, "age");
+            long sum = client.sum(null, null, User.class, "age");
             System.out.println("the sum of age column value is  :" + sum);
 
 
             // create table
-            ColumnarAdmin hbaseAdmin = constructAdmin();
+            HBaseColumnarAdmin hbaseAdmin = constructAdmin();
             hbaseAdmin.createTable("hbase_client", "t", "t1", "t2");
 
             // delete table
@@ -58,7 +56,8 @@ public class Demo {
         }
     }
 
-    private static void findObjectListByPage(ColumnarClient client) throws ColumnarClientException {
+    private static void findObjectListByPage(HBaseColumnarClient client)
+                    throws ColumnarClientException {
         PageBean<User> pageBean = new PageBean<User>() {};
         pageBean.setStartRow(Bytes.toBytes(1));
         pageBean.setStopRow(Bytes.toBytes(20));
@@ -72,20 +71,20 @@ public class Demo {
         }
     }
 
-    private static void findObjectRange(ColumnarClient client) throws ColumnarClientException {
+    private static void findObjectRange(HBaseColumnarClient client) throws ColumnarClientException {
         List<User> findObjectList =
                         client.findObjectList(Bytes.toBytes(1), Bytes.toBytes(4), User.class);
 
         System.out.println("the users count is:" + findObjectList.size());
     }
 
-    private static void findObject(ColumnarClient client) throws ColumnarClientException {
+    private static void findObject(HBaseColumnarClient client) throws ColumnarClientException {
         User findObject = client.findObject(Bytes.toBytes(1), User.class);
         System.out.println(findObject.getUserName());
         System.out.println(findObject.getId());
     }
 
-    private static void putObject(ColumnarClient client) throws ColumnarClientException {
+    private static void putObject(HBaseColumnarClient client) throws ColumnarClientException {
 
         for (int i = 1; i < 21; i++) {
             User user = new User();
@@ -97,10 +96,10 @@ public class Demo {
         }
     }
 
-    private static ColumnarClient constructClient() throws ColumnarClientException {
+    private static HBaseColumnarClient constructClient() throws ColumnarClientException {
         int scanCaching = 200;
         int scanBatch = 100;
-        ColumnarClient client = new HBaseColumnarClient(scanCaching, scanBatch);
+        HBaseColumnarClient client = new HBaseColumnarClient(scanCaching, scanBatch);
 
         DataSourceConfig config = new DataSourceConfig("hbase.properties");
 
@@ -110,8 +109,8 @@ public class Demo {
         return client;
     }
 
-    public static ColumnarAdmin constructAdmin() throws ColumnarClientException {
-        ColumnarAdmin adminClient = new HBaseColumnarAdmin();
+    public static HBaseColumnarAdmin constructAdmin() throws ColumnarClientException {
+        HBaseColumnarAdmin adminClient = new HBaseColumnarAdmin();
         DataSourceConfig config = new DataSourceConfig("hbase.properties");
         HBaseSource source = new HBaseSource(config.getProperties());
